#319 Fix Assignment Updating

+++ packtrain-service/src/main/java/edu/mines/packtrain/services/AssignmentService.java
@@ -152,6 +152,10 @@ public class AssignmentService {
     public void deleteAssignments(Set<Long> assignments, Course course) {
         log.info("Deleting {} assignments for course '{}'", assignments.size(), course.getCode());
 
+        if (assignments.isEmpty()){
+            return;
+        }
+
         assignmentRepo.deleteByCourseAndCanvasId(course, assignments);
     }
